requirements_fixed.txt
absl-py==1.3.0
adjustText==0.7.3
aeppl==0.0.38
aesara==2.8.7
alembic==1.8.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arviz==0.12.1
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
auto-sklearn==0.14.7
autokeras==1.0.20
autopage==0.5.1
backcall==0.2.0
beautifulsoup4==4.11.1
bioinfokit==2.1.0
biom-format==2.1.12
bleach==5.0.1
CacheControl==0.12.11
cachetools==5.2.0
calour==2018.5.1
certifi==2022.9.24
cffi==1.15.1
cftime==1.6.2
charset-normalizer==2.1.1
click==8.1.3
cliff==4.0.0
cloudpickle==2.2.0
cmaes==0.8.2
cmd2==2.4.2
colorlog==6.7.0
ConfigSpace==0.4.21
conorm==1.2.0
cons==0.4.5
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.32
dask==2022.10.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distributed==2022.10.0
distro==1.8.0
docrep==0.3.2
eli5==0.13.0
emcee==3.1.3
entrypoints==0.4
etuples==0.3.8
executing==1.1.1
fastjsonschema==2.16.2
fastprogress==1.0.3
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.37.4
fsspec==2022.8.2
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
greenlet==1.1.3.post0
grpcio==1.49.1
h5py==3.7.0
hdmedians==0.14.2
HeapDict==1.0.1
idna==3.4
imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==4.13.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
liac-arff==2.5.0
libclang==14.0.6
lightgbm==3.3.3
llvmlite==0.39.1
locket==1.0.0
lockfile==0.12.2
logical-unification==0.4.5
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
matplotlib-venn==0.11.7
miniKanren==1.0.3
mistune==2.0.4
msgpack==1.0.4
multipledispatch==0.6.0
mxnet==1.9.1
natsort==8.2.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
netCDF4==1.6.1
notebook==6.5.1
notebook-shim==0.1.0
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.1
opt-einsum==3.3.0
optuna==3.0.3
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prettytable==3.4.1
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDOE==0.3.8
Pygments==2.13.0
pymc==4.2.2
pynisher==0.6.4
pyparsing==3.0.9
pyperclip==1.8.2
pyrfr==0.8.3
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-bio==0.5.7
scikit-learn==0.24.2
scikit-optimize==0.9.0
scipy==1.8.1
seaborn==0.10.0
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smac==1.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
stack-data==0.5.1
statsmodels==0.13.2
stevedore==4.0.1
tabulate==0.9.0
tblib==1.7.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-cpu-aws==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.16.0
textwrap3==0.9.2
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xarray==2022.10.0
xarray-einstats==0.3.0
xgboost==1.3.3
zict==2.2.0
zipp==3.9.0